A new "we're temporarily closed" page.
This single HTML page will tell visitors that we're closed temporarily.

Due to costs in hosting (EC2), and some miscalculations on my part in terms of
Amazon IO utilization, I can't run this site 24 hours a day. I'm next working
on a plan to run this site for 8-10 hours a day, with the site being
'closed' the rest of the time, until the site can 'pay for itself'.

The HTML page uses the Skeleton 'framework':

http://www.getskeleton.com/